fasttpu: move metrics into own module
