+++ fasttpu/src/lib.rs
@@ -2,5 +2,6 @@ pub(crate) mod buf;
 pub(crate) mod cert;
 pub mod cnc;
 pub(crate) mod ip;
+pub mod metrics;
 pub(crate) mod reasm;
 pub mod server;
